LPS-55885 Moves layout revision action to a new action class
