+++ portal-impl/src/com/liferay/portlet/layoutsadmin/action/EditLayoutsAction.java
@@ -222,9 +222,6 @@ public class EditLayoutsAction extends PortletAction {
 			else if (cmd.equals("display_order")) {
 				updateDisplayOrder(actionRequest);
 			}
-			else if (cmd.equals("delete_layout_revision")) {
-				deleteLayoutRevision(actionRequest);
-			}
 			else if (cmd.equals("enable")) {
 				enableLayout(actionRequest);
 			}
@@ -250,9 +247,6 @@ public class EditLayoutsAction extends PortletAction {
 			}
 			else if (cmd.equals("select_layout_branch")) {
 				selectLayoutBranch(actionRequest);
-			}
-			else if (cmd.equals("update_layout_revision")) {
-				updateLayoutRevision(actionRequest, themeDisplay);
 			}
 
 			MultiSessionMessages.add(
@@ -469,31 +463,6 @@ public class EditLayoutsAction extends PortletAction {
 		}
 		else {
 			checkPermission(permissionChecker, group, layout, selPlid);
-		}
-	}
-
-	protected void deleteLayoutRevision(ActionRequest actionRequest)
-		throws Exception {
-
-		HttpServletRequest request = PortalUtil.getHttpServletRequest(
-			actionRequest);
-
-		long layoutRevisionId = ParamUtil.getLong(
-			actionRequest, "layoutRevisionId");
-
-		LayoutRevision layoutRevision =
-			LayoutRevisionLocalServiceUtil.getLayoutRevision(layoutRevisionId);
-
-		LayoutRevisionLocalServiceUtil.deleteLayoutRevision(layoutRevision);
-
-		boolean updateRecentLayoutRevisionId = ParamUtil.getBoolean(
-			actionRequest, "updateRecentLayoutRevisionId");
-
-		if (updateRecentLayoutRevisionId) {
-			StagingUtil.setRecentLayoutRevisionId(
-				request, layoutRevision.getLayoutSetBranchId(),
-				layoutRevision.getPlid(),
-				layoutRevision.getParentLayoutRevisionId());
 		}
 	}
 
@@ -1083,75 +1052,6 @@ public class EditLayoutsAction extends PortletAction {
 			layoutTypeSettingsProperties);
 
 		return new Object[] {layout, oldFriendlyURL};
-	}
-
-	protected void updateLayoutRevision(
-			ActionRequest actionRequest, ThemeDisplay themeDisplay)
-		throws Exception {
-
-		long layoutRevisionId = ParamUtil.getLong(
-			actionRequest, "layoutRevisionId");
-
-		LayoutRevision layoutRevision =
-			LayoutRevisionLocalServiceUtil.getLayoutRevision(layoutRevisionId);
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			actionRequest);
-
-		LayoutRevision enableLayoutRevision =
-			LayoutRevisionLocalServiceUtil.updateLayoutRevision(
-				serviceContext.getUserId(), layoutRevisionId,
-				layoutRevision.getLayoutBranchId(), layoutRevision.getName(),
-				layoutRevision.getTitle(), layoutRevision.getDescription(),
-				layoutRevision.getKeywords(), layoutRevision.getRobots(),
-				layoutRevision.getTypeSettings(), layoutRevision.getIconImage(),
-				layoutRevision.getIconImageId(), layoutRevision.getThemeId(),
-				layoutRevision.getColorSchemeId(),
-				layoutRevision.getWapThemeId(),
-				layoutRevision.getWapColorSchemeId(), layoutRevision.getCss(),
-				serviceContext);
-
-		if (layoutRevision.getStatus() != WorkflowConstants.STATUS_INCOMPLETE) {
-			StagingUtil.setRecentLayoutRevisionId(
-				themeDisplay.getUser(), layoutRevision.getLayoutSetBranchId(),
-				layoutRevision.getPlid(), layoutRevision.getLayoutRevisionId());
-
-			return;
-		}
-
-		LayoutRevision lastLayoutRevision =
-			LayoutRevisionLocalServiceUtil.fetchLastLayoutRevision(
-				enableLayoutRevision.getPlid(), true);
-
-		if (lastLayoutRevision != null) {
-			LayoutRevision newLayoutRevision =
-				LayoutRevisionLocalServiceUtil.addLayoutRevision(
-					serviceContext.getUserId(),
-					layoutRevision.getLayoutSetBranchId(),
-					layoutRevision.getLayoutBranchId(),
-					enableLayoutRevision.getLayoutRevisionId(), false,
-					layoutRevision.getPlid(),
-					lastLayoutRevision.getLayoutRevisionId(),
-					lastLayoutRevision.isPrivateLayout(),
-					lastLayoutRevision.getName(), lastLayoutRevision.getTitle(),
-					lastLayoutRevision.getDescription(),
-					lastLayoutRevision.getKeywords(),
-					lastLayoutRevision.getRobots(),
-					lastLayoutRevision.getTypeSettings(),
-					lastLayoutRevision.isIconImage(),
-					lastLayoutRevision.getIconImageId(),
-					lastLayoutRevision.getThemeId(),
-					lastLayoutRevision.getColorSchemeId(),
-					lastLayoutRevision.getWapThemeId(),
-					lastLayoutRevision.getWapColorSchemeId(),
-					lastLayoutRevision.getCss(), serviceContext);
-
-			StagingUtil.setRecentLayoutRevisionId(
-				themeDisplay.getUser(),
-				newLayoutRevision.getLayoutSetBranchId(),
-				newLayoutRevision.getPlid(),
-				newLayoutRevision.getLayoutRevisionId());
-		}
 	}
 
 	protected void updateLookAndFeel(
