+++ portal-web/docroot/html/portlet/staging_bar/view_layout_revision_details.jsp
@@ -188,7 +188,7 @@ else {
 			%>
 
 			<portlet:actionURL var="publishURL">
-				<portlet:param name="struts_action" value="/staging_bar/edit_layouts" />
+				<portlet:param name="struts_action" value="/staging_bar/edit_layout_revision" />
 				<portlet:param name="<%= Constants.CMD %>" value="update_layout_revision" />
 				<portlet:param name="redirect" value="<%= PortalUtil.getLayoutFullURL(themeDisplay) %>" />
 				<portlet:param name="layoutRevisionId" value="<%= String.valueOf(layoutRevision.getLayoutRevisionId()) %>" />
